+++ src/utils/img_detect_ai.py
@@ -48,10 +48,12 @@ def detect_img_py_v1(img_path):
                     result = "other"
                 return result + ": " + str(max_score)
             else:
-                logger.warning(f"unknown error, detail: {result}")
+                logger.warning(f"unknown error, detail: {img_path}, please re click AI detect img")
+                constants.detect_model_flag = False
                 return False
     except Exception as e:
-        logger.warning(f"unknown error, detail: {e}")
+        logger.warning(f"unknown error, detail: {e}, please re click AI detect img")
+        constants.detect_model_flag = False
         return False
 
 
@@ -155,10 +157,12 @@ def model_detect_img_java_v1(img_path):
                     result = "other"
                 return result + ": " + str(max_score)
             else:
-                logger.warning(f"unknown error, detail: {img_path}")
+                logger.warning(f"unknown error, detail: {img_path}, please re click AI detect img")
+                constants.detect_model_flag = False
                 return False
     except Exception as e:
-        logger.warning(f"unknown error, detail: {e}")
+        logger.warning(f"unknown error, detail: {e}, please re click AI detect img")
+        constants.detect_model_flag = False
         return False